refactor: move cleanURL to a utility
